requirements.txt
spotipy>=2.22.0
PyPDF2>=2.0.0
python-dotenv>=0.19.0
requests>=2.27.0